lib/requirements.txt
python-dotenv==1.0.0

pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

aiokafka==0.12.0